refactor(cli): typed progress handles, literal justify, shared pkg-root helper
- display: _RichProgressHandle now speaks rich's TaskID natively; the
  plain-text justify is a Literal; _last_desc is Optional[str]
- _helpers: _human_bytes accepts float (int callers unchanged) so the
  KB/MB divide typechecks
- uninstall: the three copy-pasted find_spec('cairn').origin lookups
  collapse into one _pkg_root() helper that handles the None module
  spec explicitly (same try/except fallback at every call site)
- validate: loop variable no longer reuses the deleted except-as name

+++ src/cairn/cli/_helpers.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 import json
 
 
-def _human_bytes(n: int) -> str:
+def _human_bytes(n: float) -> str:
     for unit in ("B", "KB", "MB", "GB"):
         if n < 1024:
             return f"{n:.0f} {unit}"
